requirements.txt
--index-url https://download.pytorch.org/whl/cu126
--extra-index-url https://pypi.org/simple
absl-py==2.1.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
antlr4-python3-runtime==4.9.3
async-timeout==5.0.1
attrs==25.1.0
bitsandbytes==0.45.2
cloudpickle==3.1.1
contourpy==1.3.1
cycler==0.12.1
docstring_parser==0.16
econml==0.16.0
filelock==3.17.0
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2025.2.0
grpcio==1.70.0
hydra-core==1.3.2
idna==3.10
imageio==2.37.0
importlib_resources==6.5.2
Jinja2==3.1.5
joblib==1.4.2
jsonargparse==4.37.0
kiwisolver==1.4.8
lazy_loader==0.4
lightgbm==4.6.0
lightning==2.5.0.post0
lightning-utilities==0.12.0
llvmlite==0.44.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
networkx==3.4.2
numba==0.61.2
numpy==2.2.3
omegaconf==2.3.0
packaging==24.2
pandas==2.2.3
patsy==1.0.1
pillow==11.1.0
propcache==0.2.1
protobuf==5.29.3
Pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytorch-lightning==2.5.0.post0
pytz==2025.2
PyYAML==6.0.2
rich==13.9.4
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
shap==0.48.0
six==1.17.0
slicer==0.0.8
sparse==0.17.0
statsmodels==0.14.4
sympy==1.13.1
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
threadpoolctl==3.5.0
tifffile==2025.5.10
torch==2.6.0
torchmetrics==1.6.1
torchvision==0.21.0
tqdm==4.67.1
triton==3.2.0
typeshed_client==2.7.0
typing_extensions==4.12.2
tzdata==2025.2
Werkzeug==3.1.3
xformers==0.0.29.post3
yarl==1.18.3
